Removed unnecessary data loss when not updating database

+++ rooster-android-app/src/main/java/com/itborci/layers/DBAdapter.java
@@ -58,11 +58,13 @@ public class DBAdapter {
         public void onUpgrade(SQLiteDatabase db, int oldVersion, 
         int newVersion) 
         {
-            Log.w(TAG, "Upgrading database from version " + oldVersion 
-                    + " to "
-                    + newVersion + ", which will destroy all old data");
-            db.execSQL("DROP TABLE IF EXISTS subjects");
-            onCreate(db);
+            if(oldVersion != newVersion) {
+                Log.w(TAG, "Upgrading database from version " + oldVersion
+                        + " to "
+                        + newVersion + ", which will destroy all old data");
+                    db.execSQL("DROP TABLE IF EXISTS subjects");
+                    onCreate(db);
+            }
         }
     }
